Finished porting DirectoryService to use mocked filesystem implementation.
